refactor: S3Uploader 추상화
