+++ src/main/kotlin/codel/member/business/MemberService.kt
@@ -1,10 +1,10 @@
 package codel.member.business
 
 import codel.member.domain.CodeImage
+import codel.member.domain.ImageUploader
 import codel.member.domain.Member
 import codel.member.domain.MemberRepository
 import codel.member.domain.Profile
-import codel.member.domain.S3Uploader
 import codel.member.presentation.request.CodeImageSavedRequest
 import codel.member.presentation.request.MemberLoginRequest
 import codel.member.presentation.request.ProfileSavedRequest
@@ -15,7 +15,7 @@ import org.springframework.stereotype.Service
 @Service
 class MemberService(
     private val memberRepository: MemberRepository,
-    private val s3Uploader: S3Uploader,
+    private val imageUploader: ImageUploader,
 ) {
     fun loginMember(request: MemberLoginRequest): MemberLoginResponse {
         val member =
@@ -47,12 +47,19 @@ class MemberService(
     }
 
     @Transactional
-    fun saveCodeImage(request: CodeImageSavedRequest) {
-        val imageFiles = request.imageFiles
-        val codeImages = CodeImage(imageFiles.map { file -> s3Uploader.uploadFile(file) })
+    fun saveCodeImage(
+        member: Member,
+        request: CodeImageSavedRequest,
+    ) {
+        val codeImage = uploadFile(request)
 
-        memberRepository.saveImagePath(codeImages)
+        memberRepository.saveImagePath(member, codeImage)
 
         // memberStatus 수정
+    }
+
+    private fun uploadFile(request: CodeImageSavedRequest): CodeImage {
+        val imageFiles = request.imageFiles
+        return CodeImage(imageFiles.map { file -> imageUploader.uploadFile(file) })
     }
 }
