+++ src/main/kotlin/codel/member/domain/ImageUploader.kt
@@ -1,0 +1,7 @@
+package codel.member.domain
+
+import org.springframework.web.multipart.MultipartFile
+
+interface ImageUploader {
+    fun uploadFile(file: MultipartFile): String
+}
